+++ lib/models/training/training_guide.dart
