+++ lib/widgets/main/mainpage_text_widget.dart
